增加线程执行时长的tpxx下时间计算功能

+++ src/main/java/com/xiaoyaozi/threadPool/ExecutorTimeBucket.java
@@ -21,7 +21,7 @@ public class ExecutorTimeBucket {
     /**
      * 桶上限
      */
-    private Long upperBound;
+    private Integer upperBound;
     /**
      * 数量
      */
